did nothing involved with image animation lol. moved timekeeping out of engine and into a dedicated class, and also integrated delta-time to actions so everything moves consistently even under stress. TODO: now that i have a constant frame of reference, use it to animate the sprites!

+++ src/actions/action.h
@@ -208,12 +208,12 @@ private:
   const int src;
   const int dst;
   const std::vector<int> c_points;
-  const int travel_time;
+  const double travel_time;
   uint32_t mv_type;
 
   bool reverse;
   int pos;
-  int elapsed_ticks;
+  double elapsed_ticks;
 };
 
 #endif
